+++ MainGUI/Resources/Changelog.md
@@ -3,6 +3,7 @@ Modnix Changelog
 # Version 2.5.4, 2020-12-30
 
 * New: Support Steam installation, detected with priority over EGS.
+* Gui: Wiki link now opens official wiki instead of fandom wiki.
 
 # Version 2.5.3, 2020-07-09